requirements.txt
aiohttp
hurry.filesize
